restcheck_pbt (restcheck v0.0.1)

Property-based testing interface for restcheck.

backend/0

-type backend() :: restcheck_backend:t().

generator/0

-type generator() :: term().

num_tests/0

-type num_tests() :: pos_integer().

output_fun/0

-type output_fun() :: fun((string(), [term()]) -> ok).

prop/0

-type prop() :: fun((term()) -> true | {false, Reason :: term()}).

property/0

-type property() :: term().

schema/0

-type schema() :: ndto:schema().

dto(Backend, Schema)

-spec dto(Backend, Schema) -> Generator
       when Backend :: backend(), Schema :: schema(), Generator :: generator().

Asks the given backend for a DTO generator from a given schema.

forall(Backend, Generators, Prop)

-spec forall(Backend, Generators, Prop) -> ForAll
          when
              Backend :: backend(),
              Generators :: [generator()],
              Prop :: prop(),
              ForAll :: property().

Wraps a forall property in the given backend format.

noshrink(Backend, Generator)

-spec noshrink(Backend, Generator) -> NoShrinkGenerator
            when Backend :: backend(), Generator :: generator(), NoShrinkGenerator :: generator().

Disables shrinking for a given generator for a given backend.

quickcheck(Backend, Property, NumTests)

-spec quickcheck(Backend, Property, NumTests) -> Result
              when
                  Backend :: backend(),
                  Property :: property(),
                  NumTests :: num_tests(),
                  Result :: true | {false, Reason},
                  Reason :: term().

Equivalent to quickcheck(Backend, Property, NumTests, fun (Format, Data) -> io:format(Format, Data) end).

quickcheck(Backend, Property, NumTests, OutputFun)

-spec quickcheck(Backend, Property, NumTests, OutputFun) -> Result
              when
                  Backend :: backend(),
                  Property :: property(),
                  NumTests :: num_tests(),
                  OutputFun :: output_fun(),
                  Result :: true | {false, Reason},
                  Reason :: term().

Runs a property-based test using the given backend.